feat: Refactor NovaSonic client and session management with new StreamSession class
