+++ web/src/streamSession.ts
@@ -1,0 +1,119 @@
+import { Buffer } from "node:buffer";
+import { DefaultAudioInputConfiguration, DefaultSystemPrompt, DefaultTextConfiguration } from "./consts";
+import { NovaSonicBidirectionalStreamClient } from "./client";
+
+export class StreamSession {
+  private audioBufferQueue: Buffer[] = [];
+  private maxQueueSize = 200; // Maximum number of audio chunks to queue
+  private isProcessingAudio = false;
+  private isActive = true;
+
+  constructor(
+    private sessionId: string,
+    private client: NovaSonicBidirectionalStreamClient
+  ) {}
+
+  // Register event handlers for this specific session
+  public onEvent(
+    eventType: string,
+    handler: (data: any) => void
+  ): StreamSession {
+    this.client.registerEventHandler(this.sessionId, eventType, handler);
+    return this; // For chaining
+  }
+
+  public async setupPromptStart(): Promise<void> {
+    this.client.setupPromptStartEvent(this.sessionId);
+  }
+
+  public async setupSystemPrompt(
+    textConfig: typeof DefaultTextConfiguration = DefaultTextConfiguration,
+    systemPromptContent: string = DefaultSystemPrompt
+  ): Promise<void> {
+    this.client.setupSystemPromptEvent(
+      this.sessionId,
+      textConfig,
+      systemPromptContent
+    );
+  }
+
+  public async setupStartAudio(
+    audioConfig: typeof DefaultAudioInputConfiguration = DefaultAudioInputConfiguration
+  ): Promise<void> {
+    this.client.setupStartAudioEvent(this.sessionId, audioConfig);
+  }
+
+  // Stream audio for this session
+  public async streamAudio(audioData: Buffer): Promise<void> {
+    // Check queue size to avoid memory issues
+    if (this.audioBufferQueue.length >= this.maxQueueSize) {
+      // Queue is full, drop oldest chunk
+      this.audioBufferQueue.shift();
+      console.log("Audio queue full, dropping oldest chunk");
+    }
+
+    // Queue the audio chunk for streaming
+    this.audioBufferQueue.push(audioData);
+    this.processAudioQueue();
+  }
+
+  // Process audio queue for continuous streaming
+  private async processAudioQueue() {
+    if (
+      this.isProcessingAudio ||
+      this.audioBufferQueue.length === 0 ||
+      !this.isActive
+    )
+      return;
+
+    this.isProcessingAudio = true;
+    try {
+      // Process all chunks in the queue, up to a reasonable limit
+      let processedChunks = 0;
+      const maxChunksPerBatch = 5; // Process max 5 chunks at a time to avoid overload
+
+      while (
+        this.audioBufferQueue.length > 0 &&
+        processedChunks < maxChunksPerBatch &&
+        this.isActive
+      ) {
+        const audioChunk = this.audioBufferQueue.shift();
+        if (audioChunk) {
+          await this.client.streamAudioChunk(this.sessionId, audioChunk);
+          processedChunks++;
+        }
+      }
+    } finally {
+      this.isProcessingAudio = false;
+
+      // If there are still items in the queue, schedule the next processing using setTimeout
+      if (this.audioBufferQueue.length > 0 && this.isActive) {
+        setTimeout(() => this.processAudioQueue(), 0);
+      }
+    }
+  }
+  // Get session ID
+  public getSessionId(): string {
+    return this.sessionId;
+  }
+
+  public async endAudioContent(): Promise<void> {
+    if (!this.isActive) return;
+    await this.client.sendContentEnd(this.sessionId);
+  }
+
+  public async endPrompt(): Promise<void> {
+    if (!this.isActive) return;
+    await this.client.sendPromptEnd(this.sessionId);
+  }
+
+  public async close(): Promise<void> {
+    if (!this.isActive) return;
+
+    this.isActive = false;
+    this.audioBufferQueue = []; // Clear any pending audio
+
+    await this.client.sendSessionEnd(this.sessionId);
+    console.log(`Session ${this.sessionId} close completed`);
+  }
+}
